# Optimize slice operation in ImmutableArray