--- a/collections/src/main/scala/strawman/collection/immutable/ImmutableArray.scala
+++ b/collections/src/main/scala/strawman/collection/immutable/ImmutableArray.scala
@@ -191,7 +191,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[T] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofRef(Arrays.copyOfRange[T](unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofRef(Arrays.copyOfRange[T](unsafeArray, lo, hi))
     }
   }
 
@@ -209,7 +212,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Byte] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofByte(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofByte(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -227,7 +233,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Short] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofShort(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofShort(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -245,7 +254,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Char] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofChar(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofChar(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -263,7 +275,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Int] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofInt(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofInt(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -281,7 +296,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Long] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofLong(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofLong(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -299,7 +317,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Float] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofFloat(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofFloat(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -317,7 +338,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Double] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofDouble(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofDouble(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -335,7 +359,10 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
     override def slice(from: Int, until: Int): ImmutableArray[Boolean] = {
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      new ofBoolean(Arrays.copyOfRange(unsafeArray, lo, hi))
+      if (lo >= hi)
+        emptyImpl
+      else
+        new ofBoolean(Arrays.copyOfRange(unsafeArray, lo, hi))
     }
   }
 
@@ -356,9 +383,13 @@ object ImmutableArray extends StrictOptimizedClassTagSeqFactory[ImmutableArray] 
       // cant use util.Arrays.copyOfRange[Unit](repr, from, until) - Unit is special and doesnt compile
       val lo = scala.math.max(from, 0)
       val hi = scala.math.min(until, length)
-      val slicedLenght = hi - lo
-      val res = new Array[Unit](slicedLenght)
-      new ofUnit(res)
+      val slicedLength = hi - lo
+      if (slicedLength <= 0)
+        emptyImpl
+      else {
+        val res = new Array[Unit](slicedLength)
+        new ofUnit(res)
+      }
     }
   }
 }