baseline performance (0a000621122dc58cccee08f41355e3bb940f0852):
```
[info] Benchmark                              (size)  Mode  Cnt        Score         Error  Units
[info] ImmutableArrayBenchmark.access_slice        0  avgt    8       38.393 ±       0.416  ns/op
[info] ImmutableArrayBenchmark.access_slice        1  avgt    8       39.797 ±       0.606  ns/op
[info] ImmutableArrayBenchmark.access_slice        2  avgt    8       38.203 ±       0.261  ns/op
[info] ImmutableArrayBenchmark.access_slice        3  avgt    8       37.194 ±       0.369  ns/op
[info] ImmutableArrayBenchmark.access_slice        4  avgt    8       41.499 ±       0.251  ns/op
[info] ImmutableArrayBenchmark.access_slice        7  avgt    8       40.420 ±       0.485  ns/op
[info] ImmutableArrayBenchmark.access_slice        8  avgt    8       37.346 ±       0.207  ns/op
[info] ImmutableArrayBenchmark.access_slice       15  avgt    8       41.066 ±       0.387  ns/op
[info] ImmutableArrayBenchmark.access_slice       16  avgt    8       38.310 ±       0.326  ns/op
[info] ImmutableArrayBenchmark.access_slice       17  avgt    8       40.797 ±       0.313  ns/op
[info] ImmutableArrayBenchmark.access_slice       39  avgt    8       42.464 ±       0.477  ns/op
[info] ImmutableArrayBenchmark.access_slice      282  avgt    8      107.206 ±       2.077  ns/op
[info] ImmutableArrayBenchmark.access_slice     4096  avgt    8     1483.075 ±      36.826  ns/op
[info] ImmutableArrayBenchmark.access_slice   131070  avgt    8    55750.166 ±    1819.028  ns/op
[info] ImmutableArrayBenchmark.access_slice  7312102  avgt    8  6388168.153 ± 7012236.507  ns/op
```

This PR improvement:
```
[info] Benchmark                              (size)  Mode  Cnt       Score       Error  Units
[info] ImmutableArrayBenchmark.access_slice        0  avgt    8       7.399 ±     0.096  ns/op
[info] ImmutableArrayBenchmark.access_slice        1  avgt    8       6.912 ±     0.084  ns/op
[info] ImmutableArrayBenchmark.access_slice        2  avgt    8       6.996 ±     0.021  ns/op
[info] ImmutableArrayBenchmark.access_slice        3  avgt    8       7.126 ±     0.078  ns/op
[info] ImmutableArrayBenchmark.access_slice        4  avgt    8       7.434 ±     0.028  ns/op
[info] ImmutableArrayBenchmark.access_slice        7  avgt    8       7.713 ±     0.063  ns/op
[info] ImmutableArrayBenchmark.access_slice        8  avgt    8       7.684 ±     0.110  ns/op
[info] ImmutableArrayBenchmark.access_slice       15  avgt    8       8.599 ±     0.100  ns/op
[info] ImmutableArrayBenchmark.access_slice       16  avgt    8       7.867 ±     0.090  ns/op
[info] ImmutableArrayBenchmark.access_slice       17  avgt    8       9.248 ±     0.097  ns/op
[info] ImmutableArrayBenchmark.access_slice       39  avgt    8      11.298 ±     0.548  ns/op
[info] ImmutableArrayBenchmark.access_slice      282  avgt    8      27.505 ±     1.063  ns/op
[info] ImmutableArrayBenchmark.access_slice     4096  avgt    8     310.161 ±    36.247  ns/op
[info] ImmutableArrayBenchmark.access_slice   131070  avgt    8    9640.269 ±   403.856  ns/op
[info] ImmutableArrayBenchmark.access_slice  7312102  avgt    8  676886.167 ± 22375.895  ns/op
```