@Ichoran 2018-01-20
Thanks for looking at this!  The behavior needs to be tweaked to work the way `slice` is supposed to, but otherwise this should be a better implementation.  The benchmarking doesn't seem to be a good indication of actual time taken for small array sizes, but the benefit is obvious on the large sizes.