@ackratos 2018-02-25
rerun benchmark against 2.12.4 with following sbt setting:

```
val collectionsScalaVersionSettings = Seq(
  scalaVersion := "2.12.4",
  crossScalaVersions := scalaVersion.value :: "2.12.4" :: dotty.value :: Nil
)

val commonSettings = Seq(
  organization := "ch.epfl.scala",
  version := "0.10.0-SNAPSHOT",
  scalaVersion := "2.12.4",
```
```
This PR improvement

[info] Benchmark                              (size)  Mode  Cnt       Score       Error  Units
[info] ImmutableArrayBenchmark.access_slice        0  avgt    8       7.478 ±     0.145  ns/op
[info] ImmutableArrayBenchmark.access_slice        1  avgt    8       7.356 ±     0.662  ns/op
[info] ImmutableArrayBenchmark.access_slice        2  avgt    8       7.189 ±     0.265  ns/op
[info] ImmutableArrayBenchmark.access_slice        3  avgt    8       7.199 ±     0.057  ns/op
[info] ImmutableArrayBenchmark.access_slice        4  avgt    8       7.357 ±     0.108  ns/op
[info] ImmutableArrayBenchmark.access_slice        7  avgt    8       7.928 ±     0.145  ns/op
[info] ImmutableArrayBenchmark.access_slice        8  avgt    8       7.893 ±     0.285  ns/op
[info] ImmutableArrayBenchmark.access_slice       15  avgt    8       8.623 ±     0.071  ns/op
[info] ImmutableArrayBenchmark.access_slice       16  avgt    8       8.013 ±     0.078  ns/op
[info] ImmutableArrayBenchmark.access_slice       17  avgt    8       8.149 ±     0.080  ns/op
[info] ImmutableArrayBenchmark.access_slice       39  avgt    8      11.391 ±     0.086  ns/op
[info] ImmutableArrayBenchmark.access_slice      282  avgt    8      28.116 ±     0.771  ns/op
[info] ImmutableArrayBenchmark.access_slice     4096  avgt    8     312.608 ±    11.314  ns/op
[info] ImmutableArrayBenchmark.access_slice   131070  avgt    8   10262.133 ±  1544.902  ns/op
[info] ImmutableArrayBenchmark.access_slice  7312102  avgt    8  688483.807 ± 27744.618  ns/op

collection strawman baseline (0a000621122dc58cccee08f41355e3bb940f0852)
[info] Benchmark                              (size)  Mode  Cnt        Score        Error  Units
[info] ImmutableArrayBenchmark.access_slice        0  avgt    8       38.762 ±      0.630  ns/op
[info] ImmutableArrayBenchmark.access_slice        1  avgt    8       40.028 ±      0.548  ns/op
[info] ImmutableArrayBenchmark.access_slice        2  avgt    8       56.638 ±     22.198  ns/op
[info] ImmutableArrayBenchmark.access_slice        3  avgt    8       41.656 ±      0.351  ns/op
[info] ImmutableArrayBenchmark.access_slice        4  avgt    8       41.610 ±      0.448  ns/op
[info] ImmutableArrayBenchmark.access_slice        7  avgt    8       37.348 ±      0.340  ns/op
[info] ImmutableArrayBenchmark.access_slice        8  avgt    8       40.937 ±      0.583  ns/op
[info] ImmutableArrayBenchmark.access_slice       15  avgt    8       40.982 ±      0.867  ns/op
[info] ImmutableArrayBenchmark.access_slice       16  avgt    8       41.430 ±      0.896  ns/op
[info] ImmutableArrayBenchmark.access_slice       17  avgt    8       38.804 ±      0.314  ns/op
[info] ImmutableArrayBenchmark.access_slice       39  avgt    8       47.292 ±      0.715  ns/op
[info] ImmutableArrayBenchmark.access_slice      282  avgt    8      112.369 ±     14.016  ns/op
[info] ImmutableArrayBenchmark.access_slice     4096  avgt    8     1475.208 ±     44.008  ns/op
[info] ImmutableArrayBenchmark.access_slice   131070  avgt    8    54926.844 ±   1527.051  ns/op
[info] ImmutableArrayBenchmark.access_slice  7312102  avgt    8  4875963.582 ± 487106.138  ns/op

scala 2.12.x official baseline (9fc13560ff91304c45d85b6a1e17f73c8fcec699)
[info] Benchmark                     (size)  Mode  Cnt        Score       Error  Units
[info] ArrayBenchmark.access_slice        0  avgt    8       40.889 ±     4.991  ns/op
[info] ArrayBenchmark.access_slice        1  avgt    8       39.278 ±     3.365  ns/op
[info] ArrayBenchmark.access_slice        2  avgt    8       37.768 ±     0.117  ns/op
[info] ArrayBenchmark.access_slice        3  avgt    8       37.608 ±     0.116  ns/op
[info] ArrayBenchmark.access_slice        4  avgt    8       37.708 ±     0.290  ns/op
[info] ArrayBenchmark.access_slice        7  avgt    8       37.578 ±     0.350  ns/op
[info] ArrayBenchmark.access_slice        8  avgt    8       37.337 ±     0.146  ns/op
[info] ArrayBenchmark.access_slice       15  avgt    8       37.009 ±     0.438  ns/op
[info] ArrayBenchmark.access_slice       16  avgt    8       36.992 ±     0.303  ns/op
[info] ArrayBenchmark.access_slice       17  avgt    8       37.094 ±     0.225  ns/op
[info] ArrayBenchmark.access_slice       39  avgt    8       36.549 ±     0.342  ns/op
[info] ArrayBenchmark.access_slice      282  avgt    8       44.495 ±     0.501  ns/op
[info] ArrayBenchmark.access_slice     4096  avgt    8      313.988 ±     9.483  ns/op
[info] ArrayBenchmark.access_slice   131070  avgt    8     9970.620 ±   327.554  ns/op
[info] ArrayBenchmark.access_slice  7312102  avgt    8  1148737.239 ± 58234.753  ns/op
```