@szeiger 2018-02-27
My previous comment still stands: Since `slice` is duplicated in all specialized implementations, it should return a specialized type. The way it is implemented now you lose specialization. With the new implementation that special-cases empty arrays this means that we need to add empty instances of all specialized types.

With the current non-specialized return types you could use a single implementation of `slice` that calls `ArrayOps.copyOf` (just merged as part of #479) but I think it's worth having the specialized return types.