@ackratos 2018-03-05
Latest benchmark (return specialized ImmutableArray to avoid auto-boxing):
```
[info] Benchmark                                    (size)  Mode  Cnt       Score       Error  Units
[info] ImmutableArrayBenchmark.access_slice              0  avgt   16       8.655 ±     0.126  ns/op
[info] ImmutableArrayBenchmark.access_slice              1  avgt   16       8.658 ±     0.069  ns/op
[info] ImmutableArrayBenchmark.access_slice              2  avgt   16       8.704 ±     0.024  ns/op
[info] ImmutableArrayBenchmark.access_slice              3  avgt   16       8.806 ±     0.093  ns/op
[info] ImmutableArrayBenchmark.access_slice              4  avgt   16       8.895 ±     0.097  ns/op
[info] ImmutableArrayBenchmark.access_slice              7  avgt   16       9.148 ±     0.033  ns/op
[info] ImmutableArrayBenchmark.access_slice              8  avgt   16       9.273 ±     0.125  ns/op
[info] ImmutableArrayBenchmark.access_slice             15  avgt   16       9.941 ±     0.108  ns/op
[info] ImmutableArrayBenchmark.access_slice             16  avgt   16       9.786 ±     0.480  ns/op
[info] ImmutableArrayBenchmark.access_slice             17  avgt   16       9.685 ±     0.042  ns/op
[info] ImmutableArrayBenchmark.access_slice             39  avgt   16      12.235 ±     0.125  ns/op
[info] ImmutableArrayBenchmark.access_slice            282  avgt   16      27.228 ±     0.252  ns/op
[info] ImmutableArrayBenchmark.access_slice           4096  avgt   16     302.324 ±     3.077  ns/op
[info] ImmutableArrayBenchmark.access_slice         131070  avgt   16    9618.849 ±   346.984  ns/op
[info] ImmutableArrayBenchmark.access_slice        7312102  avgt   16  674628.854 ± 11980.765  ns/op
```