@szeiger 2018-03-07
Could you run a benchmark against the update in https://github.com/scala/collection-strawman/pull/492? Does this version here still offer an advantage? As far as I can tell the difference comes down to 1 vs 2 or 3 operations on polymorphic arrays and any remaining advantage would be small.